+++ Lab4/src/Matrix.cpp
@@ -14,7 +14,7 @@ Vector4 Vector4::operator*(const TransformMatrix& matrix) const {
     };
 }
 
-double Vector4::dot(const Vector4& other) const {
+float Vector4::dot(const Vector4& other) const {
     return x * other.x + y * other.y + z * other.z;
 }
 
@@ -35,7 +35,7 @@ Vector4 Vector4::operator-(const Vector4& other) const {
     return {x - other.x, y - other.y, z - other.z, w - other.w};
 }
 
-Vector4 Vector4::operator*(double scalar) const {
+Vector4 Vector4::operator*(float scalar) const {
     return {x * scalar, y * scalar, z * scalar, w * scalar};
 }
 
@@ -55,7 +55,7 @@ Vector4 &Vector4::operator-=(const Vector4 &other) {
     return *this;
 }
 
-Vector4& Vector4::operator*=(double scalar) {
+Vector4& Vector4::operator*=(float scalar) {
     x *= scalar;
     y *= scalar;
     z *= scalar;
@@ -64,10 +64,10 @@ Vector4& Vector4::operator*=(double scalar) {
 }
 
 Vector4& Vector4::operator*=(const TransformMatrix& matrix) {
-    double new_x = matrix[0][0] * x + matrix[0][1] * y + matrix[0][2] * z + matrix[0][3] * w;
-    double new_y = matrix[1][0] * x + matrix[1][1] * y + matrix[1][2] * z + matrix[1][3] * w;
-    double new_z = matrix[2][0] * x + matrix[2][1] * y + matrix[2][2] * z + matrix[2][3] * w;
-    double new_w = matrix[3][0] * x + matrix[3][1] * y + matrix[3][2] * z + matrix[3][3] * w;
+    float new_x = matrix[0][0] * x + matrix[0][1] * y + matrix[0][2] * z + matrix[0][3] * w;
+    float new_y = matrix[1][0] * x + matrix[1][1] * y + matrix[1][2] * z + matrix[1][3] * w;
+    float new_z = matrix[2][0] * x + matrix[2][1] * y + matrix[2][2] * z + matrix[2][3] * w;
+    float new_w = matrix[3][0] * x + matrix[3][1] * y + matrix[3][2] * z + matrix[3][3] * w;
 
     x = new_x;
     y = new_y;
@@ -78,7 +78,7 @@ Vector4& Vector4::operator*=(const TransformMatrix& matrix) {
 }
 
 Vector4& Vector4::normalize() {
-    double length = std::sqrt(x * x + y * y + z * z);
+    float length = std::sqrt(x * x + y * y + z * z);
     if (length > 1e-6) { 
         x /= length;
         y /= length;
@@ -113,9 +113,9 @@ TransformMatrix operator*(const TransformMatrix& a, const TransformMatrix& b) {
     return result;
 }
 
-static TransformMatrix create_rotation_matrix_x(double angle) {
-    double c = std::cos(angle);
-    double s = std::sin(angle);
+static TransformMatrix create_rotation_matrix_x(float angle) {
+    float c = std::cos(angle);
+    float s = std::sin(angle);
     return {{
         {1, 0, 0, 0},
         {0, c, -s, 0},
@@ -124,9 +124,9 @@ static TransformMatrix create_rotation_matrix_x(double angle) {
     }};
 }
 
-static TransformMatrix create_rotation_matrix_y(double angle) {
-    double c = std::cos(angle);
-    double s = std::sin(angle);
+static TransformMatrix create_rotation_matrix_y(float angle) {
+    float c = std::cos(angle);
+    float s = std::sin(angle);
     return {{
         {c, 0, s, 0},
         {0, 1, 0, 0},
@@ -135,9 +135,9 @@ static TransformMatrix create_rotation_matrix_y(double angle) {
     }};
 }
 
-static TransformMatrix create_rotation_matrix_z(double angle) {
-    double c = std::cos(angle);
-    double s = std::sin(angle);
+static TransformMatrix create_rotation_matrix_z(float angle) {
+    float c = std::cos(angle);
+    float s = std::sin(angle);
     return {{
         {c, -s, 0, 0},
         {s, c, 0, 0},
@@ -176,7 +176,7 @@ TransformMatrix create_move_matrix(const Vector4& translation) {
     }};
 }
 
-TransformMatrix create_scale_matrix(double scalar){
+TransformMatrix create_scale_matrix(float scalar){
     return {{
         {scalar, 0, 0, 0},
         {0, scalar, 0, 0},
